Move ofp.v13 package to ofp for the sake of simple name
This patch renames the package "openflow/ofp.v13" to package
"openflow/ofp".
